+++ game_patch/purefaction/pf.cpp
@@ -1,4 +1,4 @@
-
+#include <cassert>
 #include <common/config/BuildConfig.h>
 #include "pf.h"
 #include "pf_packets.h"
@@ -11,66 +11,129 @@
 #include <sstream>
 #include <xlog/xlog.h>
 
-void process_pf_player_announce_packet(const void* data, size_t len, [[ maybe_unused ]] const rf::NetAddr& addr)
+static void send_pf_announce_player_packet(rf::Player* player, pf_pure_status pure_status)
 {
-    // Client-side packet
+    // Send: server -> client
+    assert(rf::is_server);
+
+    pf_player_announce_packet announce_packet;
+    announce_packet.hdr.type = static_cast<uint8_t>(pf_packet_type::announce_player);
+    announce_packet.hdr.size = sizeof(announce_packet); // should not include header size (PF bug)
+    announce_packet.version = pf_announce_player_packet_version;
+    announce_packet.player_id = player->net_data->player_id;
+    announce_packet.is_pure = static_cast<uint8_t>(pure_status);
+
+    auto player_list = SinglyLinkedList(rf::player_list);
+    for (auto& other_player : player_list) {
+        rf::net_send(other_player.net_data->addr, &announce_packet, sizeof(announce_packet));
+    }
+}
+
+static void process_pf_player_announce_packet(const void* data, size_t len, [[ maybe_unused ]] const rf::NetAddr& addr)
+{
+    // Receive: client <- server
     if (rf::is_server) {
         return;
     }
 
-    pf_player_announce_packet in_packet;
-    if (len < sizeof(in_packet)) {
+    pf_player_announce_packet announce_packet;
+    if (len < sizeof(announce_packet)) {
         xlog::trace("Invalid length in PF player_announce packet");
         return;
     }
 
-    std::memcpy(&in_packet, data, sizeof(in_packet));
+    std::memcpy(&announce_packet, data, sizeof(announce_packet));
 
-    if (in_packet.version != 2) {
+    if (announce_packet.version != pf_announce_player_packet_version) {
         xlog::trace("Invalid version in PF player_announce packet");
         return;
     }
 
-    xlog::trace("PF player_announce packet: player %u is_pure %d", in_packet.player_id, in_packet.is_pure);
+    xlog::trace("PF player_announce packet: player %u is_pure %d", announce_packet.player_id, announce_packet.is_pure);
 
-    if (in_packet.player_id == rf::local_player->net_data->player_id) {
+    if (announce_packet.player_id == rf::local_player->net_data->player_id) {
         static const char* pf_verification_status_names[] = { "none", "blue", "gold", "red" };
         auto pf_verification_status =
-            in_packet.is_pure < std::size(pf_verification_status_names)
-            ? pf_verification_status_names[in_packet.is_pure] : "unknown";
-        xlog::info("PF Verification Status: %s (%u)", pf_verification_status, in_packet.is_pure);
+            announce_packet.is_pure < std::size(pf_verification_status_names)
+            ? pf_verification_status_names[announce_packet.is_pure] : "unknown";
+        xlog::info("PF Verification Status: %s (%u)", pf_verification_status, announce_packet.is_pure);
     }
 }
 
-void process_pf_player_stats_packet(const void* data, size_t len, [[ maybe_unused ]] const rf::NetAddr& addr)
+static void send_pf_player_stats_packet(rf::Player* player)
 {
-    // Client-side packet
+    // Send: server -> client
+    assert(rf::is_server);
+
+    std::byte packet_buf[512];
+    pf_player_stats_packet stats_packet;
+    stats_packet.hdr.type = static_cast<uint8_t>(pf_packet_type::player_stats);
+    stats_packet.hdr.size = sizeof(stats_packet); // should not include header size (PF bug)
+    stats_packet.version = pf_player_stats_packet_version;
+    stats_packet.player_count = 0;
+
+    std::stringstream ss;
+    auto player_list = SinglyLinkedList{rf::player_list};
+    for (auto& current_player : player_list) {
+        auto& player_stats = *static_cast<PlayerStatsNew*>(current_player.stats);
+        pf_player_stats_packet::player_stats out_stats;
+        out_stats.player_id = current_player.net_data->player_id;
+        out_stats.is_pure = 0;
+        out_stats.accuracy = 0;
+        out_stats.streak_max = 0;
+        out_stats.streak_current = 0;
+        out_stats.kills = player_stats.num_kills;
+        out_stats.deaths = player_stats.num_deaths;
+        out_stats.team_kills = 0;
+        ++stats_packet.player_count;
+        stats_packet.hdr.size += sizeof(out_stats);
+        ss.write(reinterpret_cast<char*>(&out_stats), sizeof(out_stats));
+    }
+
+    auto stats_str = ss.str();
+    int packet_len = sizeof(stats_packet) + stats_str.size();
+    if (static_cast<size_t>(packet_len) > sizeof(packet_buf)) {
+        // One packet should be enough for 32 players (13 bytes * 32 players = 416 bytes)
+        // Handling more data would require sending multiple packets to avoid building one big UDP datagram
+        // that would cross 512 bytes limit that is usually used as maximal datagram size
+        xlog::error("PF player_stats packet is too big: %d", packet_len);
+        return;
+    }
+
+    std::memcpy(packet_buf, &stats_packet, sizeof(stats_packet));
+    std::memcpy(packet_buf + sizeof(stats_packet), stats_str.data(), stats_str.size());
+    rf::net_send(player->net_data->addr, packet_buf, packet_len);
+}
+
+static void process_pf_player_stats_packet(const void* data, size_t len, [[ maybe_unused ]] const rf::NetAddr& addr)
+{
+    // Receive: client <- server
     if (rf::is_server) {
         return;
     }
 
-    pf_player_stats_packet in_packet;
-    if (len < sizeof(in_packet)) {
+    pf_player_stats_packet stats_packet;
+    if (len < sizeof(stats_packet)) {
         xlog::trace("Invalid length in PF player_stats packet");
         return;
     }
 
-    std::memcpy(&in_packet, data, sizeof(in_packet));
+    std::memcpy(&stats_packet, data, sizeof(stats_packet));
 
-    if (in_packet.version != 2) {
+    if (stats_packet.version != pf_player_stats_packet_version) {
         xlog::trace("Invalid version in PF player_stats packet");
         return;
     }
 
-    if (len < sizeof(in_packet) + in_packet.player_count * sizeof(pf_player_stats_packet::player_stats)) {
+    if (len < sizeof(stats_packet) + stats_packet.player_count * sizeof(pf_player_stats_packet::player_stats)) {
         xlog::trace("Invalid length in pf_player_stats packet");
         return;
     }
 
-    xlog::trace("PF player_stats packet (count %d)", in_packet.player_count);
+    xlog::trace("PF player_stats packet (count %d)", stats_packet.player_count);
 
-    auto player_stats_ptr = static_cast<const std::byte*>(data) + sizeof(in_packet);
-    for (int i = 0; i < in_packet.player_count; ++i) {
+    const std::byte* player_stats_ptr = static_cast<const std::byte*>(data) + sizeof(stats_packet);
+    for (int i = 0; i < stats_packet.player_count; ++i) {
         pf_player_stats_packet::player_stats in_stats;
         std::memcpy(&in_stats, player_stats_ptr, sizeof(in_stats));
         player_stats_ptr += sizeof(in_stats);
@@ -86,11 +149,11 @@ void process_pf_player_stats_packet(const void* data, size_t len, [[ maybe_unuse
     }
 }
 
-void process_pf_players_request_packet([[ maybe_unused ]] const void* data, [[ maybe_unused ]] size_t len, const rf::NetAddr& addr)
+static void process_pf_players_request_packet([[ maybe_unused ]] const void* data, [[ maybe_unused ]] size_t len, const rf::NetAddr& addr)
 {
     xlog::trace("PF players_request packet");
 
-    // Server-side packet
+    // Receive: server <- client (can be non-joined)
     if (!rf::is_server) {
         return;
     }
@@ -102,20 +165,20 @@ void process_pf_players_request_packet([[ maybe_unused ]] const void* data, [[ m
     }
     auto players_str = ss.str();
 
-    pf_players_packet response_fixed_part;
-    response_fixed_part.hdr.type = static_cast<uint8_t>(pf_packet_type::players);
-    response_fixed_part.hdr.size = static_cast<uint16_t>(sizeof(pf_players_packet) + players_str.size() - sizeof(rf_packet_header));
-    response_fixed_part.version = 1;
-    response_fixed_part.show_ip = 0;
+    pf_players_packet players_packet;
+    players_packet.hdr.type = static_cast<uint8_t>(pf_packet_type::players);
+    players_packet.hdr.size = static_cast<uint16_t>(sizeof(pf_players_packet) + players_str.size() - sizeof(rf_packet_header));
+    players_packet.version = 1;
+    players_packet.show_ip = 0;
 
-    auto response_size = sizeof(response_fixed_part) + players_str.size();
+    auto response_size = sizeof(players_packet) + players_str.size();
     auto response_buf = std::make_unique<std::byte[]>(response_size);
-    std::memcpy(response_buf.get(), &response_fixed_part, sizeof(response_fixed_part));
-    std::memcpy(response_buf.get() + sizeof(response_fixed_part), players_str.data(), players_str.size());
+    std::memcpy(response_buf.get(), &players_packet, sizeof(players_packet));
+    std::memcpy(response_buf.get() + sizeof(players_packet), players_str.data(), players_str.size());
     rf::net_send(addr, response_buf.get(), response_size);
 }
 
-void process_pf_packet(const void* data, int len, const rf::NetAddr& addr, [[maybe_unused]] const rf::Player* player)
+void process_pf_packet(const void* data, int len, const rf::NetAddr& addr, rf::Player* player)
 {
     rf_packet_header header;
     if (len < static_cast<int>(sizeof(header))) {
@@ -131,8 +194,16 @@ void process_pf_packet(const void* data, int len, const rf::NetAddr& addr, [[may
             process_pf_server_hash_packet(data, len, addr);
             break;
 
+        case pf_packet_type::client_hash:
+            process_pf_client_hash_packet(data, len, addr, player);
+            break;
+
         case pf_packet_type::request_cheat_check:
             process_pf_request_cheat_check_packet(data, len, addr);
+            break;
+
+        case pf_packet_type::client_cheat_check:
+            process_pf_client_cheat_check_packet(data, len, player);
             break;
 
         case pf_packet_type::announce_player:
@@ -151,4 +222,17 @@ void process_pf_packet(const void* data, int len, const rf::NetAddr& addr, [[may
             // ignore
             break;
     }
+}
+
+void pf_process_new_player(rf::Player* player)
+{
+    assert(rf::is_server);
+    send_pf_server_hash_packet(player);
+}
+
+void pf_player_verified(rf::Player* player, pf_pure_status pure_status)
+{
+    assert(rf::is_server);
+    send_pf_announce_player_packet(player, pure_status);
+    send_pf_player_stats_packet(player);
 }